refactor: flatten project structure and split Background component
- Flatten components/core/* and components/misc/* into sections/ (pages)
  and components/ (reusable primitives).
- Split 437-line Background.js into thin wrapper, useBackgroundScene hook,
  and shaders.js so WebGL setup, animation loop, and React lifecycle
  are no longer tangled.
- Extract theme handling from App.js into hooks/useTheme.js.
- Move global styles into styles/ (tokens.css, global.css, layout.css)
  and drop the duplicate prefers-color-scheme block since JS already
  sets data-theme on the root element.
