+++ src/background/shaders.js
@@ -1,0 +1,140 @@
+export const vertexShader = /* glsl */ `
+  attribute vec2 position;
+  void main() {
+    gl_Position = vec4(position, 0.0, 1.0);
+  }
+`;
+
+export const fragmentShader = /* glsl */ `
+  precision highp float;
+
+  uniform vec2  iResolution;
+  uniform float iTime;
+
+  uniform float uHeight;
+  uniform float uBaseHalf;
+  uniform mat3  uRot;
+  uniform int   uUseBaseWobble;
+  uniform float uGlow;
+  uniform vec2  uOffsetPx;
+  uniform float uNoise;
+  uniform float uSaturation;
+  uniform float uScale;
+  uniform float uHueShift;
+  uniform float uColorFreq;
+  uniform float uBloom;
+  uniform float uCenterShift;
+  uniform float uInvBaseHalf;
+  uniform float uInvHeight;
+  uniform float uMinAxis;
+  uniform float uPxScale;
+  uniform float uTimeScale;
+
+  vec4 tanh4(vec4 x){
+    vec4 e2x = exp(2.0*x);
+    return (e2x - 1.0) / (e2x + 1.0);
+  }
+
+  float rand(vec2 co){
+    return fract(sin(dot(co, vec2(12.9898, 78.233))) * 43758.5453123);
+  }
+
+  float sdOctaAnisoInv(vec3 p){
+    vec3 q = vec3(abs(p.x) * uInvBaseHalf, abs(p.y) * uInvHeight, abs(p.z) * uInvBaseHalf);
+    float m = q.x + q.y + q.z - 1.0;
+    return m * uMinAxis * 0.5773502691896258;
+  }
+
+  float sdPyramidUpInv(vec3 p){
+    float oct = sdOctaAnisoInv(p);
+    float halfSpace = -p.y;
+    return max(oct, halfSpace);
+  }
+
+  mat3 hueRotation(float a){
+    float c = cos(a), s = sin(a);
+    mat3 W = mat3(
+      0.299, 0.587, 0.114,
+      0.299, 0.587, 0.114,
+      0.299, 0.587, 0.114
+    );
+    mat3 U = mat3(
+       0.701, -0.587, -0.114,
+      -0.299,  0.413, -0.114,
+      -0.300, -0.588,  0.886
+    );
+    mat3 V = mat3(
+       0.168, -0.331,  0.500,
+       0.328,  0.035, -0.500,
+      -0.497,  0.296,  0.201
+    );
+    return W + U * c + V * s;
+  }
+
+  void main(){
+    vec2 f = (gl_FragCoord.xy - 0.5 * iResolution.xy - uOffsetPx) * uPxScale;
+
+    float z = 5.0;
+    float d = 0.0;
+
+    vec3 p;
+    vec4 o = vec4(0.0);
+
+    float centerShift = uCenterShift;
+    float cf = uColorFreq;
+
+    mat2 wob = mat2(1.0);
+    if (uUseBaseWobble == 1) {
+      float t = iTime * uTimeScale;
+      float c0 = cos(t + 0.0);
+      float c1 = cos(t + 33.0);
+      float c2 = cos(t + 11.0);
+      wob = mat2(c0, c1, c2, c0);
+    }
+
+    const int STEPS = 30;
+    for (int i = 0; i < STEPS; i++) {
+      p = vec3(f, z);
+      p.xz = p.xz * wob;
+      p = uRot * p;
+      vec3 q = p;
+      q.y += centerShift;
+      d = 0.1 + 0.2 * abs(sdPyramidUpInv(q));
+      z -= d;
+      o += (sin((p.y + z) * cf + vec4(0.0, 1.0, 2.0, 3.0)) + 1.0) / d;
+    }
+
+    o = tanh4(o * o * (uGlow * uBloom) / 1e5);
+
+    vec3 col = o.rgb;
+    float n = rand(gl_FragCoord.xy + vec2(iTime));
+    col += (n - 0.5) * uNoise;
+    col = clamp(col, 0.0, 1.0);
+
+    float L = dot(col, vec3(0.2126, 0.7152, 0.0722));
+    col = clamp(mix(vec3(L), col, uSaturation), 0.0, 1.0);
+
+    if(abs(uHueShift) > 0.0001){
+      col = clamp(hueRotation(uHueShift) * col, 0.0, 1.0);
+    }
+
+    gl_FragColor = vec4(col, o.a);
+  }
+`;
+
+export function setMat3FromEuler(yawY, pitchX, rollZ, out) {
+  const cy = Math.cos(yawY), sy = Math.sin(yawY);
+  const cx = Math.cos(pitchX), sx = Math.sin(pitchX);
+  const cz = Math.cos(rollZ), sz = Math.sin(rollZ);
+
+  out[0] = cy * cz + sy * sx * sz;
+  out[1] = cx * sz;
+  out[2] = -sy * cz + cy * sx * sz;
+  out[3] = -cy * sz + sy * sx * cz;
+  out[4] = cx * cz;
+  out[5] = sy * sz + cy * sx * cz;
+  out[6] = sy * cx;
+  out[7] = -sx;
+  out[8] = cy * cx;
+  return out;
+}
