+++ src/background/useBackgroundScene.js
@@ -1,0 +1,242 @@
+import { useEffect } from 'react';
+import { Renderer, Triangle, Program, Mesh } from 'ogl';
+import { vertexShader, fragmentShader, setMat3FromEuler } from './shaders';
+
+const DPR = 0.75;
+
+export function useBackgroundScene(containerRef, options) {
+  const {
+    height,
+    baseWidth,
+    animationType,
+    glow,
+    offsetX,
+    offsetY,
+    noise,
+    transparent,
+    scale,
+    hueShift,
+    colorFrequency,
+    hoverStrength,
+    inertia,
+    bloom,
+    suspendWhenOffscreen,
+    timeScale
+  } = options;
+
+  useEffect(() => {
+    const container = containerRef.current;
+    if (!container) return;
+
+    const H = Math.max(0.001, height);
+    const BW = Math.max(0.001, baseWidth);
+    const BASE_HALF = BW * 0.5;
+    const GLOW = Math.max(0.0, glow);
+    const NOISE = Math.max(0.0, noise);
+    const SAT = transparent ? 1.5 : 1;
+    const SCALE = Math.max(0.001, scale);
+    const HUE = hueShift || 0;
+    const CFREQ = Math.max(0.0, colorFrequency || 1);
+    const BLOOM = Math.max(0.0, bloom || 1);
+    const TS = Math.max(0, timeScale || 1);
+    const HOVSTR = Math.max(0, hoverStrength || 1);
+    const INERT = Math.max(0, Math.min(1, inertia || 0.12));
+
+    const renderer = new Renderer({ dpr: DPR, alpha: transparent, antialias: false });
+    const gl = renderer.gl;
+    gl.disable(gl.DEPTH_TEST);
+    gl.disable(gl.CULL_FACE);
+    gl.disable(gl.BLEND);
+
+    Object.assign(gl.canvas.style, {
+      position: 'absolute',
+      inset: '0',
+      width: '100%',
+      height: '100%',
+      display: 'block'
+    });
+    container.appendChild(gl.canvas);
+
+    const geometry = new Triangle(gl);
+    const iResBuf = new Float32Array(2);
+    const offsetPxBuf = new Float32Array(2);
+
+    const program = new Program(gl, {
+      vertex: vertexShader,
+      fragment: fragmentShader,
+      uniforms: {
+        iResolution: { value: iResBuf },
+        iTime: { value: 0 },
+        uHeight: { value: H },
+        uBaseHalf: { value: BASE_HALF },
+        uUseBaseWobble: { value: 1 },
+        uRot: { value: new Float32Array([1, 0, 0, 0, 1, 0, 0, 0, 1]) },
+        uGlow: { value: GLOW },
+        uOffsetPx: { value: offsetPxBuf },
+        uNoise: { value: NOISE },
+        uSaturation: { value: SAT },
+        uScale: { value: SCALE },
+        uHueShift: { value: HUE },
+        uColorFreq: { value: CFREQ },
+        uBloom: { value: BLOOM },
+        uCenterShift: { value: H * 0.25 },
+        uInvBaseHalf: { value: 1 / BASE_HALF },
+        uInvHeight: { value: 1 / H },
+        uMinAxis: { value: Math.min(BASE_HALF, H) },
+        uPxScale: { value: 1 / ((gl.drawingBufferHeight || 1) * 0.1 * SCALE) },
+        uTimeScale: { value: TS }
+      }
+    });
+    const mesh = new Mesh(gl, { geometry, program });
+    const drawFrame = () => renderer.render({ scene: mesh });
+
+    const resize = () => {
+      const w = container.clientWidth || 1;
+      const h = container.clientHeight || 1;
+      renderer.setSize(w, h);
+      iResBuf[0] = gl.drawingBufferWidth;
+      iResBuf[1] = gl.drawingBufferHeight;
+      offsetPxBuf[0] = offsetX * DPR;
+      offsetPxBuf[1] = offsetY * DPR;
+      program.uniforms.uPxScale.value = 1 / ((gl.drawingBufferHeight || 1) * 0.1 * SCALE);
+      drawFrame();
+    };
+    const ro = new ResizeObserver(resize);
+    ro.observe(container);
+    resize();
+
+    const rotBuf = new Float32Array(9);
+    const NOISE_IS_ZERO = NOISE < 1e-6;
+    let raf = 0;
+    const t0 = performance.now();
+    const startRAF = () => {
+      if (raf) return;
+      raf = requestAnimationFrame(render);
+    };
+    const stopRAF = () => {
+      if (!raf) return;
+      cancelAnimationFrame(raf);
+      raf = 0;
+    };
+
+    const rnd = () => Math.random();
+    const wX = 0.3 + rnd() * 0.6;
+    const wY = 0.2 + rnd() * 0.7;
+    const wZ = 0.1 + rnd() * 0.5;
+    const phX = rnd() * Math.PI * 2;
+    const phZ = rnd() * Math.PI * 2;
+
+    let yaw = 0, pitch = 0, roll = 0;
+    let targetYaw = 0, targetPitch = 0;
+    const lerp = (a, b, t) => a + (b - a) * t;
+
+    const pointer = { x: 0, y: 0, inside: true };
+    const onMove = e => {
+      const ww = Math.max(1, window.innerWidth);
+      const wh = Math.max(1, window.innerHeight);
+      const nx = (e.clientX - ww * 0.5) / (ww * 0.5);
+      const ny = (e.clientY - wh * 0.5) / (wh * 0.5);
+      pointer.x = Math.max(-1, Math.min(1, nx));
+      pointer.y = Math.max(-1, Math.min(1, ny));
+      pointer.inside = true;
+    };
+    const onLeave = () => { pointer.inside = false; };
+    const onBlur = () => { pointer.inside = false; };
+
+    let onPointerMove = null;
+    if (animationType === 'hover') {
+      onPointerMove = e => { onMove(e); startRAF(); };
+      window.addEventListener('pointermove', onPointerMove, { passive: true });
+      window.addEventListener('mouseleave', onLeave);
+      window.addEventListener('blur', onBlur);
+      program.uniforms.uUseBaseWobble.value = 0;
+    } else if (animationType === '3drotate') {
+      program.uniforms.uUseBaseWobble.value = 0;
+    } else {
+      program.uniforms.uUseBaseWobble.value = 1;
+    }
+
+    const render = t => {
+      const time = (t - t0) * 0.001;
+      program.uniforms.iTime.value = time;
+
+      let continueRAF = true;
+
+      if (animationType === 'hover') {
+        const maxPitch = 0.6 * HOVSTR;
+        const maxYaw = 0.6 * HOVSTR;
+        targetYaw = (pointer.inside ? -pointer.x : 0) * maxYaw;
+        targetPitch = (pointer.inside ? pointer.y : 0) * maxPitch;
+        yaw = lerp(yaw, targetYaw, INERT);
+        pitch = lerp(pitch, targetPitch, INERT);
+        roll = lerp(roll, 0, 0.1);
+        program.uniforms.uRot.value = setMat3FromEuler(yaw, pitch, roll, rotBuf);
+
+        if (NOISE_IS_ZERO) {
+          const settled =
+            Math.abs(yaw - targetYaw) < 1e-4 &&
+            Math.abs(pitch - targetPitch) < 1e-4 &&
+            Math.abs(roll) < 1e-4;
+          if (settled) continueRAF = false;
+        }
+      } else if (animationType === '3drotate') {
+        const tScaled = time * TS;
+        yaw = tScaled * wY;
+        pitch = Math.sin(tScaled * wX + phX) * 0.6;
+        roll = Math.sin(tScaled * wZ + phZ) * 0.5;
+        program.uniforms.uRot.value = setMat3FromEuler(yaw, pitch, roll, rotBuf);
+        if (TS < 1e-6) continueRAF = false;
+      } else {
+        rotBuf[0] = 1; rotBuf[1] = 0; rotBuf[2] = 0;
+        rotBuf[3] = 0; rotBuf[4] = 1; rotBuf[5] = 0;
+        rotBuf[6] = 0; rotBuf[7] = 0; rotBuf[8] = 1;
+        program.uniforms.uRot.value = rotBuf;
+        if (TS < 1e-6) continueRAF = false;
+      }
+
+      drawFrame();
+      raf = continueRAF ? requestAnimationFrame(render) : 0;
+    };
+
+    let io;
+    if (suspendWhenOffscreen) {
+      io = new IntersectionObserver(entries => {
+        const vis = entries.some(e => e.isIntersecting);
+        if (vis) startRAF();
+        else stopRAF();
+      });
+      io.observe(container);
+    }
+    startRAF();
+
+    return () => {
+      stopRAF();
+      ro.disconnect();
+      if (animationType === 'hover') {
+        if (onPointerMove) window.removeEventListener('pointermove', onPointerMove);
+        window.removeEventListener('mouseleave', onLeave);
+        window.removeEventListener('blur', onBlur);
+      }
+      if (io) io.disconnect();
+      if (gl.canvas.parentElement === container) container.removeChild(gl.canvas);
+    };
+  }, [
+    containerRef,
+    height,
+    baseWidth,
+    animationType,
+    glow,
+    noise,
+    offsetX,
+    offsetY,
+    scale,
+    transparent,
+    hueShift,
+    colorFrequency,
+    timeScale,
+    hoverStrength,
+    inertia,
+    bloom,
+    suspendWhenOffscreen
+  ]);
+}
