+++ src/hooks/useTheme.js
@@ -1,0 +1,34 @@
+import { useEffect, useState } from 'react';
+
+const THEME_COLORS = {
+  dark: '#121212',
+  light: '#f5f7fb'
+};
+
+const getPreferredTheme = () => {
+  if (typeof window === 'undefined' || !window.matchMedia) return 'dark';
+  return window.matchMedia('(prefers-color-scheme: dark)').matches ? 'dark' : 'light';
+};
+
+export function useTheme() {
+  const [theme, setTheme] = useState(getPreferredTheme);
+
+  useEffect(() => {
+    const mediaQuery = window.matchMedia('(prefers-color-scheme: dark)');
+    const handleThemeChange = event => {
+      setTheme(event.matches ? 'dark' : 'light');
+    };
+    mediaQuery.addEventListener('change', handleThemeChange);
+    return () => mediaQuery.removeEventListener('change', handleThemeChange);
+  }, []);
+
+  useEffect(() => {
+    document.documentElement.dataset.theme = theme;
+    const metaThemeColor = document.querySelector('meta[name="theme-color"]');
+    if (metaThemeColor) {
+      metaThemeColor.setAttribute('content', THEME_COLORS[theme]);
+    }
+  }, [theme]);
+
+  return theme;
+}
