+++ src/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
+import './styles/tokens.css';
+import './styles/global.css';
 import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
