+++ src/styles/layout.css
@@ -1,8 +1,3 @@
-html,
-body {
-  overscroll-behavior: none;
-}
-
 .layout-wrapper {
   position: relative;
   max-width: 768px;
@@ -35,29 +30,12 @@ body {
   animation-delay: var(--content-delay, 870ms);
 }
 
-.container > *:nth-child(1) {
-  --content-delay: 870ms;
-}
-
-.container > *:nth-child(2) {
-  --content-delay: 955ms;
-}
-
-.container > *:nth-child(3) {
-  --content-delay: 1050ms;
-}
-
-.container > *:nth-child(4) {
-  --content-delay: 1165ms;
-}
-
-.container > *:nth-child(5) {
-  --content-delay: 1295ms;
-}
-
-.container > *:nth-child(6) {
-  --content-delay: 1435ms;
-}
+.container > *:nth-child(1) { --content-delay: 870ms; }
+.container > *:nth-child(2) { --content-delay: 955ms; }
+.container > *:nth-child(3) { --content-delay: 1050ms; }
+.container > *:nth-child(4) { --content-delay: 1165ms; }
+.container > *:nth-child(5) { --content-delay: 1295ms; }
+.container > *:nth-child(6) { --content-delay: 1435ms; }
 
 @keyframes panelShadowLift {
   from {
